+++ ToyRobot/ToyRobot/robot_simulator.h
@@ -8,13 +8,13 @@ public:
 	enum DIRECTION_t { NORTH, WEST, SOUTH, EAST };
 
 	robot_simulator();
-	robot_simulator(float xPosition, float yPosition, char facingDirection, bool validPos);
+	robot_simulator(int xPosition, int yPosition, char facingDirection, bool validPos);
 	~robot_simulator();
 
 
-	void set_place(float xPosition, float yPosition, char facingDirection, bool isValidPos);
-	void move_robot(float);				//for square shaped table
-	void move_robot(float, float);		// for rectangle shaped table
+	void set_place(int xPosition, int yPosition, char facingDirection, bool isValidPos);
+	void move_robot(int);				//for square shaped table
+	void move_robot(int, int);		// for rectangle shaped table
 
 	void turn(char rotationDirection);
 	void turn_left();
@@ -29,8 +29,8 @@ public:
 
 private:
 	DIRECTION_t	m_facing_direction;
-	float		m_x_coordinate;
-	float		m_y_coordinate;
+	int		m_x_coordinate;
+	int		m_y_coordinate;
 	bool		m_onTable;
