+++ ToyRobot/ToyRobot/shape_square.cpp
@@ -7,7 +7,7 @@ shape_square::shape_square()
 	this->m_length = DEFAULT_SIZE;
 }
 
-shape_square::shape_square(float length)
+shape_square::shape_square(int length)
 {
 	m_length = length;
 }
@@ -29,12 +29,12 @@ void shape_square::draw_shape()
 
 }
 
-float shape_square::get_Dimensions()
+int shape_square::get_Dimensions()
 {
 	return m_length;
 }
 
-bool shape_square::is_valid_position(float xpos, float ypos)
+bool shape_square::is_valid_position(int xpos, int ypos)
 {
 	if (xpos > 0 && xpos <= this->m_length && ypos > 0 && ypos <= this->m_length)
 		return true;
