+++ ToyRobot/ToyRobot/shape_square.h
@@ -5,20 +5,19 @@
 
 class shape_square : public shape_base
 {
-	private:
-		float m_length;
+private:
+	int m_length;
 
-	public:
-		shape_square();
-		shape_square(float length);
-		~shape_square();
+public:
+	shape_square();
+	shape_square(int length);
+	~shape_square();
 
-		void rotate_shape();
-		void draw_shape();
+	void rotate_shape();
+	void draw_shape();
 
-		float get_Dimensions();
-		bool is_valid_position(float, float);
+	int get_Dimensions();
+	bool is_valid_position(int, int);
 
-		
 };
